Update comp_radius_mid_yoke.py
This is the very first proposed change on github... ever. :-) For my problem it works now.. 

As I understand it you want to compute the middle radus of the yoke and you want to take into account  different holes that have different Rmin and Rmax. I don't see a reason for divide between inner or an external rotor because you just compare different radii. But my understandnig could be wrong.